Fixes #80; + change `Verify integrity` behavior
- verify integrity not does not check equispacing by default anymore
